requirements.txt
beautifulsoup4==4.10.0
certifi==2024.6.2
certifi==2020.6.20
dnspython==2.6.1
python_nmap==0.7.1
Requests==2.32.3
tqdm==4.66.2
